requirements.txt
streamlit
python-dotenv
matplotlib
pandas
numpy
Jinja2
openpyxl
requests>=2.31.0
tabulate>=0.9.0
black>=24.1.1
isort>=5.13.0
plotly==5.18.0
python-dateutil>=2.8.2
pandasai>=3.0.0b17
pandasai-openai
pandasai-litellm
google-generativeai
